+++ tests/test_main.py
@@ -1,8 +1,12 @@
+import socket
+from logging import WARNING
+
 import httpx
 import pytest
 
 from tests.utils import run_server
 from uvicorn.config import Config
+from uvicorn.main import run
 
 
 async def app(scope, receive, send):
@@ -26,13 +30,33 @@ async def test_return_close_header():
     )
 
 
+def _has_ipv6(host):
+    sock = None
+    has_ipv6 = False
+    if socket.has_ipv6:
+        try:
+            sock = socket.socket(socket.AF_INET6)
+            sock.bind((host, 0))
+            has_ipv6 = True
+        except Exception:
+            pass
+    if sock:
+        sock.close()
+    return has_ipv6
+
+
 @pytest.mark.anyio
 @pytest.mark.parametrize(
     "host, url",
     [
         pytest.param(None, "http://127.0.0.1:8000", id="default"),
         pytest.param("localhost", "http://127.0.0.1:8000", id="hostname"),
-        pytest.param("::1", "http://[::1]:8000", id="ipv6"),
+        pytest.param(
+            "::1",
+            "http://[::1]:8000",
+            id="ipv6",
+            marks=pytest.mark.skipif(not _has_ipv6("::1"), reason="IPV6 not enabled"),
+        ),
     ],
 )
 async def test_run(host, url):
@@ -59,3 +83,27 @@ async def test_run_reload():
         async with httpx.AsyncClient() as client:
             response = await client.get("http://127.0.0.1:8000")
     assert response.status_code == 204
+
+
+def test_run_invalid_app_config_combination(caplog: pytest.LogCaptureFixture) -> None:
+    with pytest.raises(SystemExit) as exit_exception:
+        run(app, reload=True)
+    assert exit_exception.value.code == 1
+    assert caplog.records[-1].name == "uvicorn.error"
+    assert caplog.records[-1].levelno == WARNING
+    assert caplog.records[-1].message == (
+        "You must pass the application as an import string to enable "
+        "'reload' or 'workers'."
+    )
+
+
+def test_run_startup_failure(caplog: pytest.LogCaptureFixture) -> None:
+    async def app(scope, receive, send):
+        assert scope["type"] == "lifespan"
+        message = await receive()
+        if message["type"] == "lifespan.startup":
+            raise RuntimeError("Startup failed")
+
+    with pytest.raises(SystemExit) as exit_exception:
+        run(app, lifespan="on")
+    assert exit_exception.value.code == 3