+++ uvicorn/main.py
@@ -30,6 +30,8 @@ LIFESPAN_CHOICES = click.Choice(list(LIFESPAN.keys()))
 LOOP_CHOICES = click.Choice([key for key in LOOP_SETUPS.keys() if key != "none"])
 INTERFACE_CHOICES = click.Choice(INTERFACES)
 
+STARTUP_FAILURE = 3
+
 logger = logging.getLogger("uvicorn.error")
 
 
@@ -48,7 +50,7 @@ def print_version(ctx: click.Context, param: click.Parameter, value: bool) -> No
     ctx.exit()
 
 
-@click.command()
+@click.command(context_settings={"auto_envvar_prefix": "UVICORN"})
 @click.argument("app")
 @click.option(
     "--host",
@@ -85,14 +87,17 @@ def print_version(ctx: click.Context, param: click.Parameter, value: bool) -> No
     "reload_includes",
     multiple=True,
     help="Set glob patterns to include while watching for files. Includes '*.py' "
-    "by default, which can be overridden in reload-excludes.",
+    "by default; these defaults can be overridden with `--reload-exclude`. "
+    "This option has no effect unless watchgod is installed.",
 )
 @click.option(
     "--reload-exclude",
     "reload_excludes",
     multiple=True,
     help="Set glob patterns to exclude while watching for files. Includes "
-    "'.*, .py[cod], .sw.*, ~*' by default, which can be overridden in reload-excludes.",
+    "'.*, .py[cod], .sw.*, ~*' by default; these defaults can be overridden "
+    "with `--reload-include`. This option has no effect unless watchgod is "
+    "installed.",
 )
 @click.option(
     "--reload-delay",
@@ -140,15 +145,22 @@ def print_version(ctx: click.Context, param: click.Parameter, value: bool) -> No
 @click.option(
     "--ws-ping-interval",
     type=float,
-    default=20,
+    default=20.0,
     help="WebSocket ping interval",
     show_default=True,
 )
 @click.option(
     "--ws-ping-timeout",
     type=float,
-    default=20,
+    default=20.0,
     help="WebSocket ping timeout",
+    show_default=True,
+)
+@click.option(
+    "--ws-per-message-deflate",
+    type=bool,
+    default=True,
+    help="WebSocket per-message-deflate compression",
     show_default=True,
 )
 @click.option(
@@ -276,14 +288,14 @@ def print_version(ctx: click.Context, param: click.Parameter, value: bool) -> No
 @click.option(
     "--ssl-version",
     type=int,
-    default=SSL_PROTOCOL_VERSION,
+    default=int(SSL_PROTOCOL_VERSION),
     help="SSL version to use (see stdlib ssl module's)",
     show_default=True,
 )
 @click.option(
     "--ssl-cert-reqs",
     type=int,
-    default=ssl.CERT_NONE,
+    default=int(ssl.CERT_NONE),
     help="Whether client certificate is required (see stdlib ssl module's)",
     show_default=True,
 )
@@ -317,7 +329,6 @@ def print_version(ctx: click.Context, param: click.Parameter, value: bool) -> No
 )
 @click.option(
     "--app-dir",
-    "app_dir",
     default=".",
     show_default=True,
     help="Look for APP in the specified directory, by adding this to the PYTHONPATH."
@@ -342,6 +353,7 @@ def main(
     ws_max_size: int,
     ws_ping_interval: float,
     ws_ping_timeout: float,
+    ws_per_message_deflate: bool,
     lifespan: str,
     interface: str,
     debug: bool,
@@ -376,8 +388,6 @@ def main(
     app_dir: str,
     factory: bool,
 ) -> None:
-    sys.path.insert(0, app_dir)
-
     kwargs = {
         "host": host,
         "port": port,
@@ -389,6 +399,7 @@ def main(
         "ws_max_size": ws_max_size,
         "ws_ping_interval": ws_ping_interval,
         "ws_ping_timeout": ws_ping_timeout,
+        "ws_per_message_deflate": ws_per_message_deflate,
         "lifespan": lifespan,
         "env_file": env_file,
         "log_config": LOGGING_CONFIG if log_config is None else log_config,
@@ -421,11 +432,16 @@ def main(
         "headers": [header.split(":", 1) for header in headers],
         "use_colors": use_colors,
         "factory": factory,
+        "app_dir": app_dir,
     }
     run(app, **kwargs)
 
 
 def run(app: typing.Union[ASGIApplication, str], **kwargs: typing.Any) -> None:
+    app_dir = kwargs.pop("app_dir", None)
+    if app_dir is not None:
+        sys.path.insert(0, app_dir)
+
     config = Config(app, **kwargs)
     server = Server(config=config)
 
@@ -446,8 +462,11 @@ def run(app: typing.Union[ASGIApplication, str], **kwargs: typing.Any) -> None:
     else:
         server.run()
     if config.uds:
-        os.remove(config.uds)
+        os.remove(config.uds)  # pragma: py-win32
+
+    if not server.started and not config.should_reload and config.workers == 1:
+        sys.exit(STARTUP_FAILURE)
 
 
 if __name__ == "__main__":
-    main()
+    main()  # pragma: no cover